+++ app/src/main/java/com/example/adapp/Nav_activity.kt
@@ -5,7 +5,11 @@ import android.os.Bundle
 import android.view.View
 import androidx.fragment.app.Fragment
 import androidx.fragment.app.FragmentTransaction
-import com.google.android.material.bottomnavigation.BottomNavigationView
+import com.example.adapp.all_ads.AllAdsFragment
+import com.example.adapp.my_account.MyAccountFragment
+import com.example.adapp.my_ads.MyAdsFragment
+import com.example.adapp.new_ad.NewAdFragment
+import com.example.adapp.notification.NotificationFragment
 import kotlinx.android.synthetic.main.activity_nav_activity.*
 
 class Nav_activity : AppCompatActivity() {
@@ -49,7 +53,7 @@ class Nav_activity : AppCompatActivity() {
     //common function to set fragment to constraint layout
     private fun setCurrentFragment(fragment: Fragment)=
         supportFragmentManager.beginTransaction().apply {
-            replace(R.id.constraintLayout,fragment)
+            replace(R.id.fragment,fragment)
             commit()
         }
     fun addAdvertisement(view:View){
