+++ app/src/main/java/com/example/adapp/notification/NotificationFragment.kt
@@ -1,4 +1,4 @@
-package com.example.adapp
+package com.example.adapp.notification
 
 import android.os.Bundle
 import androidx.fragment.app.Fragment
@@ -8,6 +8,7 @@ import androidx.recyclerview.widget.RecyclerView
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
+import com.example.adapp.R
 import com.example.adapp.dummy.DummyContent
 
 /**
