+++ app/src/main/java/com/example/adapp/notification/NotificationRecyclerViewAdapter.kt
@@ -1,10 +1,11 @@
-package com.example.adapp
+package com.example.adapp.notification
 
 import androidx.recyclerview.widget.RecyclerView
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import android.widget.TextView
+import com.example.adapp.R
 
 import com.example.adapp.dummy.DummyContent.DummyItem
 
@@ -12,13 +13,13 @@ import com.example.adapp.dummy.DummyContent.DummyItem
  * [RecyclerView.Adapter] that can display a [DummyItem].
  * TODO: Replace the implementation with code for your data type.
  */
-class MyAdsRecyclerViewAdapter(
+class NotificationRecyclerViewAdapter(
     private val values: List<DummyItem>
-) : RecyclerView.Adapter<MyAdsRecyclerViewAdapter.ViewHolder>() {
+) : RecyclerView.Adapter<NotificationRecyclerViewAdapter.ViewHolder>() {
 
     override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): ViewHolder {
         val view = LayoutInflater.from(parent.context)
-            .inflate(R.layout.fragment_my_ads, parent, false)
+            .inflate(R.layout.fragment_notification, parent, false)
         return ViewHolder(view)
     }
