+++ app/src/main/res/layout/fragment_item_all_ads.xml
@@ -9,5 +9,5 @@
     android:layout_marginLeft="16dp"
     android:layout_marginRight="16dp"
     app:layoutManager="LinearLayoutManager"
-    tools:context=".AllAdsFragment"
+    tools:context=".all_ads.AllAdsFragment"
     tools:listitem="@layout/fragment_all_ads" />
