+++ app/src/main/res/layout/fragment_item_my_ads.xml
@@ -9,5 +9,5 @@
     android:layout_marginLeft="16dp"
     android:layout_marginRight="16dp"
     app:layoutManager="LinearLayoutManager"
-    tools:context=".MyAdsFragment"
+    tools:context=".my_ads.MyAdsFragment"
     tools:listitem="@layout/fragment_my_ads" />
