+++ app/src/main/res/layout/fragment_item_notification.xml
@@ -9,5 +9,5 @@
     android:layout_marginLeft="16dp"
     android:layout_marginRight="16dp"
     app:layoutManager="LinearLayoutManager"
-    tools:context=".NotificationFragment"
+    tools:context=".notification.NotificationFragment"
     tools:listitem="@layout/fragment_notification" />
