+++ app/src/main/res/layout/fragment_my_account.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".MyAccountFragment">
+    tools:context=".my_account.MyAccountFragment">
 
     <!-- TODO: Update blank fragment layout -->
     <TextView
