+++ app/src/main/res/layout/fragment_new_ad.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".NewAdFragment">
+    tools:context=".new_ad.NewAdFragment">
 
     <!-- TODO: Update blank fragment layout -->
     <TextView
